Fix OWLAP-95 Ensure modified published axioms take effect.
This fixes a bug where axioms in the delta where not always replacing the old version of the axiom from the snapshot. This was because we are building a set of current axiom objects from the OWL API which are not aware of the SNOMED axiom id so the axioms were not naturally replacing each other. We now track the axiom id independently so we can remove old axiom versions when a new one is loaded.